+++ config/application.rb
@@ -37,8 +37,6 @@ module Roda
       g.orm :active_record, primary_key_type: :uuid
     end
 
-    config.eager_load_paths += %W[#{config.root}/lib]
-
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 6.0